don't add hosts and services of a contactgroups contactgroup_members to the contactgroup itself
If a contactgroup has a contactgroup_members attribute the services
and hosts of the contactgroups in that attribute are added to the
list of services and hosts for the contactgroup in the ObjectRelations.
Therefore all get_effective..() methods that relate to contactgroups return
wrong results. This patch fixes that.

Issue #140