copy legacy model tests to Model2 test class
The Model2 class uses FakeNagiosEnvironment for config isolation.
This commit copies the legacy model tests from the Model class to
the new Model2 class.

The legacy tests are left in place if any regressions in the tests
might show up when using FNE. After a while the old tests should be
deleted and the Model2 class should probably be renamed to Model.