Parsers/LivestatusQuery - More helper methods.
Added a few helper methods to LivestatusQuery and changed
__init__() so it no longer depends on pynag.Utils.grep_to_livestatus() internally.

This is part of a bigger change to refactor and clean up adagios.utils.get* methods and support cleaner bridge
between livestatus and querystrings.

This patch and a cleanup of adagios.utils.get_services() is required before
we can proberly support opinkerfi/adagios#470

Please review and give feedback.