Consistency: Terminates every LiveStatusQuery instance with an empty line.
This also allow queries to be "directly" concatenated : query1 + query2.

Always ends a LiveStatusQuery instance by a supplementary '\n',
so to have the query finished by an empty line on its own ('\n\n').
This will allow queries to be enqueued or concatened, just directly.
And this should also be better from an API point of view because the livestatus server actually uses that to detect the end of a query when it's reading from its input/command socket/communication canal (whatever it might be) (and explicit is better than implicit also).