Updated send_command to send timestamp as an integer
When forcing a service recheck in Adagios I discovered that the command is sending the timestamp as a decimal.  The documention (http://old.nagios.org/developerinfo/externalcommands/commandinfo.php?command_id=129) shows the use of an unsigned integer and the argument description in `send_command` show it being an integer.  The issue is that Naemon 0.9.1 expects a ulong and returns `Parse error: Invalid characters (.78) in ulong '1426878227.78'.  This along with https://github.com/opinkerfi/adagios/pull/535 fixes it.