prevent invalid email or authors in `GitRepo`
Use sensible default author information if the `GitRepo` is passed
whitespace strings for `author_name` or `author_email`.

This helps solve a problem in Adagios where it uses the `REMOTE_USER`
HTTP server variable as the author, and that might be an empty string
depending on how the web server authentication is setup. I debated
changing this in Adagios, but since pynag is already trying to correct
invalid input, I figured it'd be okay for pynag to do a little bit more
input correction.