Inerhit indentation of object blocks in config
When pynag updates values it likes to use a predefined indentation schema.  This change will default to the legacy schema for new objects.  Modifications of existing objects will auto-detect and use the indentation of the previous line in the block.  This should also clean up Adagios edits, which utilize this pynag feature.